HomeServerSolution/modules/FilterModule/requirements.txt
azure-iot-device~=2.0.0
numpy